Add an emergency "critical" fan temperature.
This will still slowly ramp up to the critical fan speed, but doesn't
nicely ramp down. The goal is just that if the temperature reaches a
certain point - way, way above the setpoint - then the fans can
activate more to keep the system alive. This works nicely with having
this program set up with a fairly low set_point, and low regular fan
speed.